more progress on Issue #9
